pyproject.toml
[tool.poetry]
name = "nebula-feed"
version = "0.1.0"
description = ""
authors = ["tomfutago <tomfutago@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.0"
iconsdk = "^1.3.5"
discord-webhook = "^0.14.0"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
